:zap: Fix: 쿼리+커맨드 하나의 서비스로 작성
