+++ majorLink/src/main/java/com/example/majorLink/service/ReviewService/ReviewServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.majorLink.service.ReviewService;
 
+import com.example.majorLink.domain.Lecture;
 import com.example.majorLink.domain.Review;
 import com.example.majorLink.dto.request.ReviewRequestDTO;
 import com.example.majorLink.repository.LectureRepository;
@@ -7,6 +8,8 @@ import com.example.majorLink.repository.ReviewRepository;
 import com.example.majorLink.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -49,5 +52,15 @@ public class ReviewServiceImpl implements ReviewService{
         Review review = reviewRepository.findById(reviewId).get();
 
         reviewRepository.delete(review);
+    }
+
+    // 리뷰 페이징
+    @Override
+    public Page<Review> getReviewList(Long lectureId, Integer page) {
+        Lecture lecture = lectureRepository.findById(lectureId).get();
+
+        Page<Review> reviewPage = reviewRepository.findAllByLecture(lecture, PageRequest.of(page, 10));
+
+        return reviewPage;
     }
 }
